Revisit the alltoall benchmark mode
This patch revisits the communication pattern for the alltoall
benchmark mode.
Previously, the benchmark was divided into rounds and for each round,
each MPI rank was sending data to a rank Y and was receiving data
from rank Z.

The main issue with this pattern is that is makes difficult to
isolate which MPI ranks are slower than the others, and this is
because each rank was communicating with 2 remote ranks at any step.

This patch revisits the logic by allowing each rank to communicate
with a single remote rank at any time. This makes things much easier
when it comes the time to figure out the slow MPI ranks.

The algorithm to select the list of remote ranks to communicate with
has been inspired from the 'linktest' tool from FZ Julich:
http://www.fz-juelich.de/ias/jsc/EN/Expertise/Support/Software/LinkTest/_node.html